% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot a Boxplot}
\usage{
plot_boxplot(x, group)
}
\arguments{
\item{x}{A numeric vector.}

\item{group}{A factor or grouping variable.}
}
\value{
A ggplot2 boxplot object.
}
\description{
This function creates a boxplot of a numeric variable grouped by a factor.
}
\examples{
plot_boxplot(mtcars$mpg, mtcars$cyl)
}
