% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_density}
\alias{plot_density}
\title{Plot a Density Curve}
\usage{
plot_density(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A ggplot2 density plot object.
}
\description{
This function plots a density estimate of a numeric variable.
}
\examples{
plot_density(mtcars$mpg)
}
