% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot a Scatterplot}
\usage{
plot_scatter(x, y)
}
\arguments{
\item{x}{A numeric vector for the x-axis.}

\item{y}{A numeric vector for the y-axis.}
}
\value{
A ggplot2 scatterplot object.
}
\description{
This function creates a scatterplot of two numeric vectors using ggplot2.
}
\examples{
plot_scatter(mtcars$mpg, mtcars$hp)
}
